+++ CHANGELOG.md
@@ -53,6 +53,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [UNRELEASED]
 - Update documentation
+- Add `cancel(request)` functionality to cancel specific requests
 
 ## [0.8.4] - 2019-22-07
 - Included response data in a http error (can allow clients to display server messages)